backend/requirements.txt
Flask
flask-cors
werkzeug
gunicorn
psycopg2-binary